rename args to arguments package
