+++ flowkit/CHANGELOG.md
@@ -162,7 +162,7 @@ The `TransactionAddressesRoles` was moved to `transactions` package and renamed 
 
 ---
 
-The `ParseArgumentsJSON` and `ParseArgumentsWithoutType` were moved to a `args` package and renamed 
+The `ParseArgumentsJSON` and `ParseArgumentsWithoutType` were moved to a `arguments` package and renamed 
 to `ParseJSON` and `ParseWithoutType` correspondingly.
